% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.create_empty_table}
\alias{.create_empty_table}
\title{.create_empty_table}
\usage{
.create_empty_table(num_rows, num_cols)
}
\arguments{
\item{num_cols}{}
}
\value{
Empty data.frame object
}
\description{
Create empty data.frame. Source: https://stackoverflow.com/questions/10689055/create-an-empty-data-frame
}
\keyword{internal}
