% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.my_transpose}
\alias{.my_transpose}
\title{.my_transpose}
\usage{
.my_transpose(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
The same dataframe but transposed keeping column/row names
}
\description{
Transpose a data.frame
}
\keyword{internal}
