% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pcgcodemaker}
\alias{.pcgcodemaker}
\title{Create factor from taxonomy data. Deprecated in v4. Entero == 1; Pseudo == 2; others == 3}
\usage{
.pcgcodemaker(tax, f_ = FALSE)
}
\arguments{
\item{f_}{}
}
\value{
Factor with levels Entero, Pseudo and others
}
\description{
Create factor from taxonomy data. Deprecated in v4. Entero == 1; Pseudo == 2; others == 3
}
\keyword{internal}
