% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pcgcodemaker2}
\alias{.pcgcodemaker2}
\title{.pcgcodemaker2}
\usage{
.pcgcodemaker2(tax, together = TRUE)
}
\arguments{
\item{f_}{}
}
\value{
Factor with levels 1-4 (Bulkholderiales, Pseudomonadales, others, Xanthomonadales)
}
\description{
Create factor from taxonomy data. Deprecated in v4. Bulkholderiales, Pseudomonadales, others, Xanthomonadales
}
\keyword{internal}
