% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_growth}
\alias{full_growth}
\title{This function consists in a loop that runs growth(...)() functions as many
times as needed to reach a given population size.}
\usage{
full_growth(
  this_timestep,
  grow_step,
  abun_total,
  func = "growth",
  interactions = NULL,
  carrying_capacities = NULL
)
}
\description{
This function consists in a loop that runs growth(...)() functions as many
times as needed to reach a given population size.
}
