% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abundance_and_tax_from_table.R
\name{get_abundance_and_tax_from_table}
\alias{get_abundance_and_tax_from_table}
\title{get_abundance_and_tax_from_table}
\usage{
get_abundance_and_tax_from_table(
  exp_file,
  species_are_rows = TRUE,
  sep = "\\t",
  row.names = 1,
  skip = 0,
  taxa_fields = NULL,
  tax_col_name = "taxonomy",
  tax_sep = ";",
  NA_option = "___",
  check.names = FALSE
)
}
\arguments{
\item{exp_file}{.csv/.tsv abundance table; required}

\item{tax_sep}{4 chars or longer "breaks" renamer (will include this as if it were a real name)}

\item{check.names}{}
}
\value{
two separate data.frames: abundances and taxa data
}
\description{
This function reads a .csv/.tsv abundance table, where the last
column (or row) is the taxonomy
}
