% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{growth_log}
\alias{growth_log}
\title{This function simulates growth in a community by looking at the carrying
capacities of each species, which they have in common with other species in
their group. It takes a named vector, carrying_capacities.}
\usage{
growth_log(x, carrying_capacities)
}
\description{
Growth is RANDOM but logistic: all species have the opportunity to grow each
time this function is called, but whether they will is random and depends on
their abundance (random), but they grow at different rates depending not
only on which group they belong to but also to how close they are to their
carrying capacity  (logistic)
}
\details{
If the carrying capacity for a group was surpassed already, the species of
that group will die at a proportionate rate, while the others may grow.
}
