% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{growth_per_group}
\alias{growth_per_group}
\title{This function simulates growth in a community by looking at the carrying
capacities of the group they belong to. It takes a named vector,
carrying_capacities.}
\usage{
growth_per_group(x, carrying_capacities, grow_step, interactions = NULL)
}
\description{
Growth is RANDOM: all species have the opportunity to grow each
time this function is called, but whether they will is random and depends on
their abundance (random). They grow at different rates depending not
only on which group they belong to but also on their carrying capacity.
}
\details{
As opposed to growth(), the total growth rate is defined by the sum of the
growth rates of every group, which are defined as
[grow_step * group's \% of total carrying capacity]. Every group will have
grow_step (by default 1) of its members grow in each run, and they can be
from the same species or not. Also, how much will each species grow is
proportional to the carrying capacity of its group. This is to avoid group
extinction and also to ensure growth has a similar, proportional rate for
each group so all groups reach their CC at the same time.
}
