% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundVectorPreservingSum.R
\name{roundVectorPreservingSum}
\alias{roundVectorPreservingSum}
\title{roundVectorPreservingSum}
\usage{
roundVectorPreservingSum(vector)
}
\arguments{
\item{vector}{}
}
\value{

}
\description{
This function rounds the input vector and then calculates the difference in
the sum of the original vector and the rounded vector. To fix this difference,
it randomly selects elements from the rounded vector and adjusts their values
to preserve the sum of the original vector.
}
\details{
This is needed to avoid decimals in the final simulated communities when
simulating multiple functional groups.
}
