% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_timeseries.R
\name{simulate_timeseries}
\alias{simulate_timeseries}
\title{Main function. Simulates the changes in abundances for a given initial
abundance data. The model has a dilution and growth system. After each
dilution, a random organism is duplicated}
\usage{
simulate_timeseries(
  counts_data,
  carrying_capacities = NULL,
  dilution = 8 * 10^(-3),
  no_of_dil = 12,
  fixation_at = 1,
  abun_total = NULL,
  grow_step = 1,
  keep_all_timesteps = FALSE,
  force_continue = FALSE
)
}
\arguments{
\item{abun_total}{Total abundance for the community to grow in each dilution-growth
cycle. When this value is reached, the cycle ends and another dilution (if needed)
happens. If NULL, \code{abun_total} is set to the initial total abundance}

\item{grow_step}{Number of individuals that grow each timestep (Default: 1)}

\item{force_continue}{}
}
\value{

}
\description{
Uses c++ functions for speed
}
