% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_timeseries_with_sub.R
\name{simulate_timeseries_with_sub}
\alias{simulate_timeseries_with_sub}
\title{OLD version for the main function. Does not use c++ functions. No
\code{fixation_at} parameter. Accepts \code{growth_rate}.}
\usage{
simulate_timeseries_with_sub(
  counts_data,
  dilution = 8 * 10^(-3),
  no_of_dil = 12,
  N = NULL,
  abun_total = NULL,
  grow_step = 1,
  subs_step = 1,
  keep_all_timesteps = FALSE
)
}
\arguments{
\item{N}{Substitutions happen N times each before each dilution, N being
by default the initial total abundance}

\item{abun_total}{Total abundance for the community to grow in each dilution-growth
cycle. When this value is reached, the cycle ends and another dilution (if needed)
happens. If NULL, \code{abun_total} is set to the initial total abundance}

\item{grow_step}{Number of individuals that grow each timestep (Default: 1)}

\item{subs_step}{Number of individuals substituted each time step}
}
\value{

}
\description{
More importantly, this model is more directly based on the UNTB and has
substitutions
}
\details{
This function simulates the changes in abundances for a given initial
abundance data. The model is based on the UNTB. Each iteration, individuals
die and are randomly substituted for new ones (without immigration or
speciation).
Additionally, the model has a dilution and growth system. After each
dilution, a random organism is duplicated. This happens, by default, until
the total abundance reaches the value previous to the dilution. But it can
be changed with the option abun_total
}
