% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundVectorPreservingSum.R
\name{roundVectorPreservingSum}
\alias{roundVectorPreservingSum}
\title{roundVectorPreservingSum}
\usage{
roundVectorPreservingSum(vector, carrying_capacities = NULL)
}
\arguments{
\item{vector}{}
}
\value{

}
\description{
This function rounds the input vector and then calculates the difference in
the sum of the original vector and the rounded vector. To fix this difference,
it randomly selects elements from the rounded vector and adjusts their values
to preserve the sum of the original vector.
}
\details{
If carrying_capacities is not null, this is done in a group by group fashion.

This function is tailored for this specific application, and our abundance
vectors will always sum an integer. But in the case it was given a vector
which sum totals a decimal number (i. e. 20.2), the resulting vector will not
have the same total value but a rounded one (i. e. 20)]

This is needed to avoid decimals in the final simulated communities when
simulating multiple functional groups.
}
